+++ cli/v-i18n/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/veypi/vigo/logv"
 )
 
-var version = "v0.8.2"
+var version = "v0.8.3"
 
 // 全局配置参数
 var globalOpts = struct {