+++ ui/env.js
@@ -1,8 +1,6 @@
-
-/* 初始化模块变量池
- * $mod.$axios: axios 对象
- * $mod.$bus: 模块事件总线对象
- * $mod.scoped: 模块根路径
- * */
 export default async ($mod) => {
+  console.log($mod)
+  const res = await fetch(`${$mod.scoped}/langs.json`)
+  const messages = await res.json()
+  $mod.$i18n.load(messages)
 }